[#1] Move reference operations into overlay adapter
Generator no longer generatates `isXxxReference` and `getXxxReference`
for refable properties, and in fact the whole notion of refables is
gone. The `Overlay` adapter class has been expanded to make it possible
to target any overlay object without too much difficulty, and there are
now `isReference` and `getReference` methods defined for it. Each method
takes a parameter that identifies a child of whatever overlay the
adapter is adapting:

* A field name for a `PropertiesOverlay` (name of the internal member
  name of the implementation class)
* A value key for `MapOverlay`
* An index for `ListOverlay`

The reference methods of the overlay will provide information about how
that child is specified in the model, whether inlined or via reference.