Implement consistent value constructor semantics
Value based constructors now always create deep (with respect to
JsonOverlays) copies of the provided values, rather than using the
supplied value object as-is. This makes a difference for lists, maps,
and properties overlays. E.g. in a ListOverlay<Foo>, where Foo extends
PropertyOverlay, all the Foo values will be recreated, and so will their
property values, etc.

When reference support is added, this behavior will be altered so that
references to objects scope of the value will be copied as-is, while
internal references will be rewritten to refer to corresponding objects
within the copy.

There's a new interal method, _copy(), which all JsonOverlay objects
implement, which creates a copy using the value-based constructor.