Add _elaborateFully() method
This is important when parsed structures, or their substructures, may be
used in ways that depend on their equals semantics. When the parsed
structure is not fully elaborated, later elaboration can change hash
codes for the parsed structure or any of its pieces.

Of course, any intentional modification of the structures will also
affect equality, and could interfere with things like retrieval from
hash maps, etc. But that's up the programmer. Here we're providing a
means of ensuring that the structure will not change as a result of read
operations.