[#27] Integrate position info into parsed result
Functionality is now present. Location information gathered from the
parse are made available by a new Overlay wrapper method,
`getPositionInfo()` (and convenience method `getStartPosition()` when
you don't care about the endpoint.

Character position within the containing document, line number, and
column number are all provided. Char position is zero-based, others are
1-based.

Position of a reference is that of the reference node. Position of the
resolved value is that of the referenced node. New Overlay method
`getRefereenceOverlay()` is added to obtain the JsonOverlay object that
is the resolved target of a reference, in order to provide access to
this position info.

A number of bugs in the initial position-aware parser implementation are
fixed, here. Tests are provided for the YAML parser, but not for the
JSON parser, and the latter has not been exercised at all and probably
has bugs.

Also, SnakeYaml preprocessing of model text is incompatible with this
implementation, so for now that preprocessing is suppressed. Options
will be provided in the future to control this, and perhaps something
less drastic than suppressing position information can be worked
out. E.g. rendering the SnakeYaml output and then parsing with Jackson,
collecting position info. That position info probably won't reflect
positions in the original source file, but perhaps it's better than
nothing? It's a headache for later...