[#27] Always use YAML parser to read model files
It turns out that the Jackson JSON parser provides extremely broken
token location information in many versions, including the one we're
currently using. As of v2.1, JSON is a true subset of YAML (modulo
treatment of duplicate keys in an object - an error in YAML but silently
ignored with unspecified results in JSON). So in order to have reliable
position info, we'll use the Jackson YAML parser for both JSON and YAML
source files.